api: add RuntimeHook gRPC call (1/2): definitions
This is intended for runtime hooks to pass information to the agent.
In this patch we add a CreateContainer call so that the agent can be
notified about the creation of a new container.

- cgroupPath: cgroup path
- rootDir: root directory of the container
- annotations: annotations for the container see https://github.com/opencontainers/runtime-spec/blob/main/config.md#annotations

This patch modifies the proto definition and introduces stub functions
The next patch will generate the necessary code from the proto file.

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>