bpf/{kprobes,traceoints}: add policy filter check
This patch adds a policy filter check in the generic kprobes and
tracepoints. This is done via a hashmap of hashmaps (policy_filter_maps)
that is indexed by policy id. The resulting map is a hashmap that
includes every cgroup id that matches this policy.

This patch uses the cgroup id as retrieved in exec time. Upcoming
patches will use an appropriate cgroup tracker id.

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>