tracing: move GenericKprobeObjs to pkg/kernel
This function returns the object files for generic kprobes based on the
underlying kernel. We move it to pkg/kernel, so that it can be used by
the user-space policy filter map (pkg/policyfilter) in subsequent
patches. It will be used to generate the map based on the spec found in
the object file.

We move it to pkg/kernel to avoid import cycles.

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>