pkg/sensors: introduce policyHandler
policyHandler is the new version for SpecHandler. The difference is that
instead of an interface{} that is typically the spec itself, it gets a
TracingPolicy argument as well as the tracing policy id, which can be
used to set up policy filtering.

Subsequent patches will make use of the PolicyHandler to implement
filtering on the tracing sensor (generic kprobes and tracepoints).

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>