pkg/config: preparation for parsing changes
This patch refactors pkg/config, which is where the code for parsing
yaml files is kept. This is currently used for tracing policies, so that
we can load the YAML files directly from the tetra cli or via providing
a filename to the agent.

The main change is that the functions will now return an interface of
tracingpolicy.TracingPolicy. This allows us to have them return a
namespaced tracing policy in subsequent patches.

The tracingpolicy.TracingPolicyNamespaced, as well as the modifications
to the parsing code will be added in subsequent patches.

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>