confmap: do not fail if map does not exist
This patch modifies the confmap code to not fail if the map does not
exist. Instead, the map will be created.

While we are at it, it cleanusp up the confmap code:
 * removes uses of the now deprecated pkg/bpf
 * removes code duplication, and uses a single function for setting the
   value of the confmap.

This allows is us to run the code that initialized the config map
earlier in the tettragon init process.

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>