pkg/cgroups/fsscan: cgroup ids by scanning the fs
This patch adds the pkg/cgroup/fsscan package, that can be used to
retrieve cgroup ids based on container names. This will be used in a
subsequent patch, for implementing the pod watcher handliers for
policyfilter.

To avoid scanning the full filesystem, we keep a list of pod directories
(e.g., group/memory/kubepods/burstable). If a container id is not found
in the existing list, but it was found later, we add the pod directory
so that subsequent scans can use it.

There are also some cases where the container-id was found under a
directory that does not match the pod id. We add a special error for
this cases, so that callers will be able to distinguish this case.

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>