pkg/policyfilter: add pod event handlers
This patch adds pod event handlers for policyfilter.

A new function RegisterPodHandlers is exported, which allows callers to
register the needed hooks.

The hooks monitor pod additions and deletions, and update the
policyfilter accordingly.

Removing containers is easy, we just remove the state for a given
container id.

Adding containers is more involved because we do not the cgroup id to
enter in the bpf map. If a run-time hook was already executed, our state
should be up-to-date, so there is no more work todo. However, there are
cases where this might not have happend. One example of this is when the
agent starts, and the existing pods are processed.

In this case, we need to find the cgroup id from the container name. To
do so, we use cgroups.fsscan. Hence, the function AddPodContainer is
modified to take a pointer to a cgroup id. If that pointer is not-nil,
everything works as before. Otherwise, the code will try to determine
the cgroup id by scanning the cgroup fs.

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>