Increase the number of values in matchBinaries
Before that we had a limit to 4 values in the matchBinaries selector.
This patch uses a map per kprobe (sel_names_map) to remove this
limitation. The current limit is 256 values (the size of the map) and
should be enough for all cases.

As a follow-up we can also clear entries from the (shared) names_map
when we remove kprobes. For now we also increase the size of that map to
256 entries. This means that we can define up to 256 unique binary names
among all matchBinaries selectors.

Signed-off-by: Anastasios Papagiannis <tasos.papagiannnis@gmail.com>